Update scripts and dependencies in package.json
This commit modifies two scripts in the infrastructure-supabase package.json file, specifically "supabase:gen-types" and "supabase:reset", to directly use supabase command instead of bunx. In addition, the 'supabase' dependency was moved from 'dependencies' to 'peerDependencies'. The change ensures that the 'supabase' dependency is trusted at all levels of the project.